+++ README.md
@@ -1,2 +1,9 @@
 # pr_UFRU_SE
 student project in software engineering
+
+## Тесты
+
+* Запус тестов в локальном окружении:
+```bash
+./script/run_tests.sh
+```
