+++ image_object_detector/lib/singleton.py
@@ -1,0 +1,7 @@
+class Singleton(type):
+  instance = None
+
+  def __call__(cls, *args, **kwargs):
+    if cls.instance is None:
+      cls.instance = super(Singleton, cls).__call__(*args, **kwargs)
+    return cls.instance
